Server/requirements.txt
azure-identity==1.15.0
azure-monitor-ingestion==1.0.3
